api/src/llms/tools/fetchWebPage.tool/tool.ts
import type { JSX } from 'preact';
import LLMTool from 'api/llms/llmTool.ts';
import type { LLMToolInputSchema, LLMToolRunResult } from 'api/llms/llmTool.ts';
import {
	formatToolResult as formatToolResultBrowser,
	formatToolUse as formatToolUseBrowser,
} from './formatter.browser.tsx';
import {
	formatToolResult as formatToolResultConsole,
	formatToolUse as formatToolUseConsole,
} from './formatter.console.ts';
import type LLMConversationInteraction from 'api/llms/conversationInteraction.ts';
import type { ConversationLogEntryContentToolResult } from 'shared/types.ts';
import type { LLMAnswerToolUse } from 'api/llms/llmMessage.ts';
import FetchManager from 'shared/fetchManager.ts';
import type ProjectEditor from 'api/editor/projectEditor.ts';
import { logger } from 'shared/logger.ts';

export default class LLMToolFetchWebPage extends LLMTool {
	get inputSchema(): LLMToolInputSchema {
		return {
			type: 'object',
			properties: {
				url: { type: 'string', description: 'The URL of the web page to fetch' },
			},
			required: ['url'],
		};
	}

	formatToolUse(toolInput: LLMToolInputSchema, format: 'console' | 'browser'): string | JSX.Element {
		return format === 'console' ? formatToolUseConsole(toolInput) : formatToolUseBrowser(toolInput);
	}

	formatToolResult(
		resultContent: ConversationLogEntryContentToolResult,
		format: 'console' | 'browser',
	): string | JSX.Element {
		return format === 'console' ? formatToolResultConsole(resultContent) : formatToolResultBrowser(resultContent);
	}

	async runTool(
		_interaction: LLMConversationInteraction,
		toolUse: LLMAnswerToolUse,
		_projectEditor: ProjectEditor,
	): Promise<LLMToolRunResult> {
		const { url } = toolUse.toolInput as { url: string };
		try {
			const fetchManager = await new FetchManager().init();
			const content: string = await fetchManager.fetchPage(url);

			return {
				toolResults: this.extractTextFromHtml(content),
				toolResponse: `Successfully fetched content from ${url}`,
				bbaiResponse: {
					data: {
						url,
						html: content,
					},
				},
			};
		} catch (error) {
			logger.error(`Failed to fetch web page: ${error.message}`);

			const toolResults = `⚠️  ${error.message}`;
			const bbaiResponse = `BBai failed to fetch web page. Error: ${error.message}`;
			const toolResponse = `Failed to fetch web page. Error: ${error.message}`;
			return { toolResults, toolResponse, bbaiResponse };
		}
	}

	private extractTextFromHtml(html: string): string {
		// This is a simple extraction method. In the future, we can implement more sophisticated parsing.
		return html.replace(/<[^>]*>/g, ' ')
			.replace(/\s+/g, ' ')
			.trim();
	}
}
